move llm loading to utils
